+++ src/PostgREST/Query/Statements.hs
@@ -1,28 +1,15 @@
 {-|
 Module      : PostgREST.Query.Statements
-Description : PostgREST single SQL statements.
-
-This module constructs single SQL statements that can be parametrized and prepared.
-
-- It consumes the SqlQuery types generated by the QueryBuilder module.
-- It generates the body format and some headers of the final HTTP response.
+Description : PostgREST SQL statements.
 -}
 module PostgREST.Query.Statements
   ( prepareWrite
   , prepareRead
   , prepareCall
   , preparePlanRows
-  , ResultSet (..)
   ) where
 
-import qualified Data.Aeson.Lens                   as L
-import qualified Data.ByteString.Char8             as BS
-import qualified Hasql.Decoders                    as HD
-import qualified Hasql.DynamicStatements.Snippet   as SQL
-import qualified Hasql.DynamicStatements.Statement as SQL
-import qualified Hasql.Statement                   as SQL
-
-import Control.Lens ((^?))
+import qualified Hasql.DynamicStatements.Snippet as SQL
 
 import PostgREST.ApiRequest.Preferences
 import PostgREST.MediaType              (MTVndPlanFormat (..),
@@ -33,34 +20,10 @@ import PostgREST.SchemaCache.Routine    (MediaHandler (..), Routine,
 
 import Protolude
 
--- | Standard result set format used for all queries
-data ResultSet
-  = RSStandard
-  { rsTableTotal :: Maybe Int64
-  -- ^ count of all the table rows
-  , rsQueryTotal :: Int64
-  -- ^ count of the query rows
-  , rsLocation   :: [(BS.ByteString, BS.ByteString)]
-  -- ^ The Location header(only used for inserts) is represented as a list of strings containing
-  -- variable bindings like @"k1=eq.42"@, or the empty list if there is no location header.
-  , rsBody       :: BS.ByteString
-  -- ^ the aggregated body of the query
-  , rsGucHeaders :: Maybe BS.ByteString
-  -- ^ the HTTP headers to be added to the response
-  , rsGucStatus  :: Maybe Text
-  -- ^ the HTTP status to be added to the response
-  , rsInserted   :: Maybe Int64
-  -- ^ the number of rows inserted (Only used for upserts)
-  }
-  | RSPlan BS.ByteString -- ^ the plan of the query
-
-
 prepareWrite :: SQL.Snippet -> SQL.Snippet -> Bool -> Bool -> MediaType -> MediaHandler ->
-                Maybe PreferRepresentation -> Maybe PreferResolution -> [Text] -> Bool -> (SQL.Statement () ResultSet, ByteString)
-prepareWrite selectQuery mutateQuery isInsert isPut mt handler rep resolution pKeys prepared =
- (result, sql)
+                Maybe PreferRepresentation -> Maybe PreferResolution -> [Text] -> SQL.Snippet
+prepareWrite selectQuery mutateQuery isInsert isPut mt handler rep resolution pKeys = mtSnippet mt snippet
  where
-  result@(SQL.Statement sql _ _ _) = SQL.dynamicallyParameterized (mtSnippet mt snippet) decodeIt prepared
   checkUpsert snip = if isInsert && (isPut || resolution == Just MergeDuplicates) then snip else "''"
   pgrstInsertedF = checkUpsert "nullif(current_setting('pgrst.inserted', true),'')::int"
   snippet =
@@ -89,16 +52,9 @@ prepareWrite selectQuery mutateQuery isInsert isPut mt handler rep resolution pK
     | handler == NoAgg = "SELECT * FROM " <> sourceCTE
     | otherwise        = selectQuery
 
-  decodeIt :: HD.Result ResultSet
-  decodeIt = case mt of
-    MTVndPlan{} -> planRow
-    _           -> fromMaybe (RSStandard Nothing 0 mempty mempty Nothing Nothing Nothing) <$> HD.rowMaybe (standardRow False)
-
-prepareRead :: SQL.Snippet -> SQL.Snippet -> Bool -> MediaType -> MediaHandler -> Bool -> (SQL.Statement () ResultSet, ByteString)
-prepareRead selectQuery countQuery countTotal mt handler prepared =
- (result, sql)
+prepareRead :: SQL.Snippet -> SQL.Snippet -> Bool -> MediaType -> MediaHandler -> SQL.Snippet
+prepareRead selectQuery countQuery countTotal mt handler = mtSnippet mt snippet
  where
-  result@(SQL.Statement sql _ _ _) = SQL.dynamicallyParameterized (mtSnippet mt snippet) decodeIt prepared
   snippet =
     "WITH " <> sourceCTE <> " AS ( " <> selectQuery <> " ) " <>
     countCTEF <> " " <>
@@ -113,18 +69,11 @@ prepareRead selectQuery countQuery countTotal mt handler prepared =
 
   (countCTEF, countResultF) = countF countQuery countTotal
 
-  decodeIt :: HD.Result ResultSet
-  decodeIt = case mt of
-    MTVndPlan{} -> planRow
-    _           -> HD.singleRow $ standardRow True
 
 prepareCall :: Routine -> SQL.Snippet -> SQL.Snippet -> SQL.Snippet -> Bool ->
-               MediaType -> MediaHandler -> Bool ->
-               (SQL.Statement () ResultSet, ByteString)
-prepareCall rout callProcQuery selectQuery countQuery countTotal mt handler prepared =
-  (result, sql)
+               MediaType -> MediaHandler -> SQL.Snippet
+prepareCall rout callProcQuery selectQuery countQuery countTotal mt handler = mtSnippet mt snippet
   where
-    result@(SQL.Statement sql _ _ _) = SQL.dynamicallyParameterized (mtSnippet mt snippet) decodeIt prepared
     snippet =
       "WITH " <> sourceCTE <> " AS (" <> callProcQuery <> ") " <>
       countCTEF <>
@@ -141,49 +90,11 @@ prepareCall rout callProcQuery selectQuery countQuery countTotal mt handler prep
 
     (countCTEF, countResultF) = countF countQuery countTotal
 
-    decodeIt :: HD.Result ResultSet
-    decodeIt = case mt of
-      MTVndPlan{} -> planRow
-      _           -> fromMaybe (RSStandard (Just 0) 0 mempty mempty Nothing Nothing Nothing) <$> HD.rowMaybe (standardRow True)
 
-preparePlanRows :: SQL.Snippet -> Bool -> SQL.Statement () (Maybe Int64)
-preparePlanRows countQuery =
-  SQL.dynamicallyParameterized snippet decodeIt
-  where
-    snippet = explainF PlanJSON mempty countQuery
-    decodeIt :: HD.Result (Maybe Int64)
-    decodeIt =
-      let row = HD.singleRow $ column HD.bytea in
-      (^? L.nth 0 . L.key "Plan" .  L.key "Plan Rows" . L._Integral) <$> row
-
-standardRow :: Bool -> HD.Row ResultSet
-standardRow noLocation =
-  RSStandard <$> nullableColumn HD.int8 <*> column HD.int8
-             <*> (if noLocation then pure mempty else fmap splitKeyValue <$> arrayColumn HD.bytea)
-             <*> (fromMaybe mempty <$> nullableColumn HD.bytea)
-             <*> nullableColumn HD.bytea
-             <*> nullableColumn HD.text
-             <*> nullableColumn HD.int8
-  where
-    splitKeyValue :: ByteString -> (ByteString, ByteString)
-    splitKeyValue kv =
-      let (k, v) = BS.break (== '=') kv in
-      (k, BS.tail v)
+preparePlanRows :: SQL.Snippet -> SQL.Snippet
+preparePlanRows = explainF PlanJSON mempty
 
 mtSnippet :: MediaType -> SQL.Snippet -> SQL.Snippet
 mtSnippet mediaType snippet = case mediaType of
   MTVndPlan _ fmt opts -> explainF fmt opts snippet
   _                    -> snippet
-
--- | We use rowList because when doing EXPLAIN (FORMAT TEXT), the result comes as many rows. FORMAT JSON comes as one.
-planRow :: HD.Result ResultSet
-planRow = RSPlan . BS.unlines <$> HD.rowList (column HD.bytea)
-
-column :: HD.Value a -> HD.Row a
-column = HD.column . HD.nonNullable
-
-nullableColumn :: HD.Value a -> HD.Row (Maybe a)
-nullableColumn = HD.column . HD.nullable
-
-arrayColumn :: HD.Value a -> HD.Row [a]
-arrayColumn = column . HD.listArray . HD.nonNullable
